requirements.txt
langchain_openai==0.0.8
matplotlib==3.5.1
numpy==1.21.5
opencv_python_headless==4.9.0.80
Pillow==9.0.1
pygame==2.5.0
pytest==8.1.1
Requests==2.31.0
rich==13.7.1
scipy==1.8.0
sounddevice==0.4.6
stability_sdk==0.8.5